dtm: PoC of slot history.
The patch is a draft of a new type of history -- slot.
Also, the patch introduces a new kind of histories that connects
servers together (nexus).

These two changes open a way to DTX0-like transactions
where a transaction is beign resent between the participants
(redo) until it becomes persistent on a selected number
of nodes with persistent storage.

NOTE1: This patch introduces only a new way to share slots.
The original code allows sharing histories only between
two participants (client and server). The client could
re-integrate a DTM operation on the server if the server
looses its volatile state.
The patch provides the ability to share slots between
the servers too. This new kind of sharing is required
to make re-integration work between the servers even
if the client lost its volatile state.

NOTE2: But this kind of re-integration is not implemented
in this patch.

NOTE3: This patch violates Motr coding style and
basic programming guidelines. Consider it just as
a demo of the idea but not the right implementation
of the idea.