DTM: Introduce intent slot history.
Problem: DTM code does not have a history type
that is able to describe the semantic of intent slot.
It is impossible to comply with the DTM0 requirements without
intent slot support.

Solution: Introduce slot history type.

NOTE1: The patch does not address the other improvements
that are required for DTM0. For example, slot updates
are not sent to the remote nodes. It will be addressed
separately. The patch just opens a way to these fixes.